requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
postal==1.1.9
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2019.3
requests==2.22.0
six==1.13.0
twilio==6.35.1
urllib3==1.25.7
virtualenv==16.7.9
Werkzeug==1.0.1
